refactor: update file structure
update dependencies and refactoring
