+++ Pipfile
@@ -5,8 +5,18 @@ name = "pypi"
 
 [packages]
 loguru = "*"
+pandas = "*"
+torch = "*"
+transformers = "*"
+tqdm = "*"
+jupyter = "*"
+torchaudio = "*"
 
 [dev-packages]
+black = "*"
 
 [requires]
 python_version = "3.8"
+
+[pipenv]
+allow_prereleases = true
