+++ src/main.py
@@ -2,8 +2,10 @@ import sys
 
 from loguru import logger
 
+from model.nscm_test import test
+
 if __name__ == '__main__':
     logger.remove()
     logger.add(sys.stderr, level="INFO")
 
-    logger.info(f'hello python{3.8}')
+    test()
